functions/requirements.txt
firebase_functions~=0.1.0
openai~=1.0.0
requests~=2.31.0
python-dotenv~=1.0.0
httpx[http2]~=0.24.1
traceback2~=1.4.0
flask>=2.1.2